refactor: change import from character constants
