+++ src/mappers/map-symbols.ts
@@ -1,4 +1,4 @@
-import { characters } from '@src/constants/characters'
+import { characters } from '@src/constants'
 
 const mapSymbols = (
   useSymbols: boolean,
